third_party/WebKit/LayoutTests/fast/dom/custom/callback-context.html

<!DOCTYPE html>
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="test-harness-utils.js"></script>
<body>
<script>
(function () {

var test = async_test('callbacks created in the outer frame\'s context are ' +
                      'invoked in that context');

withFrame(function (frame) {
    var outerDocument = document;
    var proto = Object.create(frame.contentWindow.HTMLElement.prototype);
    proto.createdCallback = test.step_func(function () {
        assert_equals(document, outerDocument,
                      'the context should be the outer window');
    });
    var doc = frame.contentDocument;
    doc.registerElement('x-a', {prototype: proto});
    var script = doc.createElement('script');
    script.textContent = 'document.body.innerHTML = "<x-a></x-a>";';
    doc.body.appendChild(script);
    frame.remove();
    test.done();
});

})();


(function () {

var test = async_test('callbacks created in the inner frame\'s context are ' +
                      'invoked in that context');

withFrame(function (frame) {
    document.name = 'outer document';

    var doc = frame.contentDocument;
    doc.name = 'inner document';

    var step = test.step_func(function (event) {
        switch (event.data.step) {
        case 'remove':
            doc.body.firstChild.remove();
            break;

        case 'assert':
            window.removeEventListener('message', step);
            assert_equals(event.data.name, 'inner document',
                          'the context should be the inner window');
            frame.remove();
            test.done();

        default:
            assert_not_reached();
        };
    });
    window.addEventListener('message', step);

    var script = doc.createElement('script');
    script.textContent =
        'var proto = Object.create(HTMLElement.prototype); ' +
        'proto.detachedCallback = function () { ' +
        '    window.parent.postMessage(' +
        '        {step: "assert", name: document.name}, "*");' +
        '}; ' +
        'document.registerElement("x-a", {prototype: proto}); ' +
        'document.body.innerHTML = "<x-a></x-a>"; ' +
        'window.parent.postMessage({step: "remove"}, "*");';
    doc.body.appendChild(script);
});

})();
</script>
